update: break homepage into sections for ratings, achievements, founder introduction, Spanish learning, and call-to-action
